data-science/requirements.txt
astunparse==1.6.3
clang==5.0
click==8.0.3
flatbuffers==1.12
gast==0.4.0
google-pasta==0.2.0
h5py==3.1.0
ipykernel==6.4.1
keras==2.6.0
keras-preprocessing==1.1.2
matplotlib==3.4.3
opt-einsum==3.3.0
pandas==1.3.4
pip-chill==1.0.1
scikit-learn==1.0
tensorboard==2.7.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
wrapt==1.12.1
